Move tests next to files
