+++ src/infrastructure/yaml/storage_test.go
@@ -4,8 +4,8 @@ import (
 	"io/fs"
 	"os"
 	. "phopper/src/domain/model"
+	. "phopper/src/domain/utils"
 	. "phopper/src/infrastructure/yaml"
-	. "phopper/test/utils"
 	"slices"
 	"strings"
 	"testing"
